requirements.txt
anyio==3.6.1
asttokens==2.0.8
backcall==0.2.0
certifi==2022.9.14
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
decorator==5.1.1
executing==1.0.0
fastapi==0.82.0
h11==0.12.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
idna==3.3
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
PyJWT==2.4.0
python-dotenv==0.21.0
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
stack-data==0.5.0
starlette==0.19.1
traitlets==5.3.0
typing_extensions==4.3.0
uvicorn==0.18.3
uvloop==0.16.0
watchfiles==0.16.1
wcwidth==0.2.5
websockets==10.3
